refactor(discord): split subagent progress state
