+++ extensions/discord/src/subagent-progress.ts
@@ -1,39 +1,47 @@
 // Discord plugin module maps portable subagent progress onto source-message feedback.
-import { DEFAULT_EMOJIS } from "openclaw/plugin-sdk/channel-feedback";
-import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
-import type { PluginStateKeyedStore } from "openclaw/plugin-sdk/plugin-state-runtime";
-import {
-  normalizeOptionalLowercaseString,
-  normalizeOptionalStringifiedId,
-} from "openclaw/plugin-sdk/string-coerce-runtime";
 import { resolveDiscordAccount } from "./accounts.js";
 import { reactMessageDiscord, removeReactionDiscord } from "./send.reactions.js";
 import { sendTypingDiscord } from "./send.typing.js";
+import {
+  FAILURE_EMOJI,
+  RUNNING_EMOJIS,
+  reactionsAreAvailable,
+  reservedReactionEmojis,
+  resolveDiscordProgressTarget,
+  type DiscordProgressRequester,
+} from "./subagent-progress-config.js";
+import {
+  MAX_TRACKED_RUNS,
+  PROGRESS_STORE_TTL_MS,
+  consumeProgressRun,
+  getProgressStore,
+  listProgressStateForKey,
+  logFailure,
+  lookupProgressRun,
+  markProgressRunForCleanup,
+  persistedProgressRunFromTracker,
+  persistProgressRun,
+  resetDiscordSubagentProgressStateForTest,
+  runQueued,
+  type PersistedProgressRun,
+  type PersistProgressResult,
+  type ProgressApi,
+  type ProgressTracker,
+} from "./subagent-progress-state.js";
 
-const RUNNING_EMOJIS = ["1️⃣", "2️⃣", "3️⃣", "4️⃣", "5️⃣", "6️⃣", "7️⃣", "8️⃣", "9️⃣", "🔟"];
-const FAILURE_EMOJI = "🔴";
 const TYPING_INTERVAL_MS = 8_500;
 const TYPING_TTL_MS = 60 * 60_000;
-const PROGRESS_STORE_TTL_MS = 7 * 24 * 60 * 60_000;
 const TERMINAL_TOMBSTONE_TTL_MS = 60 * 60_000;
 const TERMINAL_LOOKUP_RETRY_MS = 1_000;
 const TERMINAL_RETRY_MAX_DELAY_MS = 60 * 60_000;
 const TERMINAL_RETRY_MAX_ATTEMPTS = 12;
 const STARTUP_RETRY_MAX_ATTEMPTS = 12;
-const MAX_TRACKED_RUNS = 4_096;
 
 type SubagentProgressEvent =
   | {
       phase: "started";
       runId: string;
-      requester?: {
-        channel?: string;
-        accountId?: string;
-        to?: string;
-        threadId?: string | number;
-        channelId?: string | number;
-        messageId?: string | number;
-      };
+      requester?: DiscordProgressRequester;
     }
   | {
       phase: "ended";
@@ -41,55 +49,6 @@ type SubagentProgressEvent =
       outcome: "ok" | "error" | "timeout" | "killed" | "unknown";
       requester?: Extract<SubagentProgressEvent, { phase: "started" }>["requester"];
     };
-
-type ProgressApi = {
-  config: OpenClawConfig;
-  logger: { debug?: (message: string) => void };
-  runtime?: {
-    state: {
-      openKeyedStore<T>(options: {
-        namespace: string;
-        maxEntries: number;
-        overflowPolicy: "reject-new";
-        defaultTtlMs: number;
-      }): PluginStateKeyedStore<T>;
-    };
-  };
-};
-
-type PersistedProgressRun = {
-  /** Plugin SQLite ownership lets an ended run clean reactions after gateway restart. */
-  key: string;
-  accountId: string;
-  channelId: string;
-  messageId: string;
-  status: "active" | "cleanup";
-  runningEmoji?: string;
-};
-
-type ProgressTracker = {
-  accountId: string;
-  channelId: string;
-  messageId: string;
-  activeRunIds: Set<string>;
-  persistedRunIds: Set<string>;
-  runningEmoji?: string;
-  runningEmojiConfirmed: boolean;
-  reactionsEnabled: boolean;
-  typingTimer?: ReturnType<typeof setInterval>;
-  typingExpiresAt: number;
-};
-
-type ProgressStateForKeyResult =
-  | { ok: true; activeRunIds: string[]; cleanupRunIds: string[]; ownedEmojis: string[] }
-  | { ok: false };
-
-type ProgressRunLookupResult =
-  | { status: "found"; value: PersistedProgressRun }
-  | { status: "missing" }
-  | { status: "error" };
-
-type PersistProgressResult = "persisted" | "terminal" | "conflict" | "error";
 
 type PersistedReconciliationResult =
   | { ok: false }
@@ -103,7 +62,6 @@ type PersistedReconciliationResult =
 
 const trackers = new Map<string, ProgressTracker>();
 const trackerKeyByRunId = new Map<string, string>();
-const trackerQueues = new Map<string, Promise<void>>();
 const terminalRunIds = new Map<string, number>();
 const terminalRetryTimers = new Map<string, ReturnType<typeof setTimeout>>();
 const terminalRetryExpiresAt = new Map<string, number>();
@@ -112,58 +70,6 @@ const startupRecoveryRetries = new Map<
   ProgressApi,
   { attempts: number; timer?: ReturnType<typeof setTimeout> }
 >();
-let progressStores = new WeakMap<object, PluginStateKeyedStore<PersistedProgressRun> | null>();
-
-function channelIdFromTarget(target?: string): string | undefined {
-  const trimmed = target?.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  if (trimmed.startsWith("channel:")) {
-    return trimmed.slice("channel:".length).trim() || undefined;
-  }
-  return /^\d+$/u.test(trimmed) ? trimmed : undefined;
-}
-
-function resolveTarget(event: Extract<SubagentProgressEvent, { phase: "started" }>) {
-  if (normalizeOptionalLowercaseString(event.requester?.channel) !== "discord") {
-    return undefined;
-  }
-  const channelId =
-    normalizeOptionalStringifiedId(event.requester?.channelId) ??
-    channelIdFromTarget(event.requester?.to);
-  const messageId = normalizeOptionalStringifiedId(event.requester?.messageId);
-  if (!channelId || !messageId) {
-    return undefined;
-  }
-  return { channelId, messageId };
-}
-
-function reservedReactionEmojis(config: OpenClawConfig, ackReaction?: string): Set<string> {
-  const reserved = new Set<string>(Object.values(DEFAULT_EMOJIS));
-  for (const emoji of Object.values(config.messages?.statusReactions?.emojis ?? {})) {
-    if (emoji?.trim()) {
-      reserved.add(emoji.trim());
-    }
-  }
-  for (const emoji of [config.messages?.ackReaction, ackReaction]) {
-    if (emoji?.trim()) {
-      reserved.add(emoji.trim());
-    }
-  }
-  for (const agent of config.agents?.list ?? []) {
-    const emoji = agent.identity?.emoji?.trim();
-    if (emoji) {
-      reserved.add(emoji);
-    }
-  }
-  return reserved;
-}
-
-function reactionsAreAvailable(config: OpenClawConfig, ackReaction?: string): boolean {
-  const reserved = reservedReactionEmojis(config, ackReaction);
-  return !RUNNING_EMOJIS.some((emoji) => reserved.has(emoji)) && !reserved.has(FAILURE_EMOJI);
-}
 
 function markRunTerminal(runId: string) {
   const now = Date.now();
@@ -209,159 +115,6 @@ function cancelTerminalRetryTimer(runId: string) {
     clearTimeout(timer);
   }
   terminalRetryTimers.delete(runId);
-}
-
-function getProgressStore(
-  api: ProgressApi,
-): PluginStateKeyedStore<PersistedProgressRun> | undefined {
-  const cached = progressStores.get(api);
-  if (cached !== undefined) {
-    return cached ?? undefined;
-  }
-  if (!api.runtime?.state) {
-    progressStores.set(api, null);
-    return undefined;
-  }
-  try {
-    const store = api.runtime.state.openKeyedStore<PersistedProgressRun>({
-      namespace: "subagent-progress",
-      maxEntries: MAX_TRACKED_RUNS,
-      overflowPolicy: "reject-new",
-      defaultTtlMs: PROGRESS_STORE_TTL_MS,
-    });
-    progressStores.set(api, store);
-    return store;
-  } catch (error) {
-    logFailure(api, "state store open", error);
-    return undefined;
-  }
-}
-
-async function persistProgressRun(
-  api: ProgressApi,
-  runId: string,
-  tracker: ProgressTracker,
-): Promise<PersistProgressResult> {
-  const store = getProgressStore(api);
-  if (!store) {
-    return "error";
-  }
-  const value = persistedProgressRunFromTracker(tracker, "active");
-  try {
-    if (await store.registerIfAbsent(runId, value)) {
-      return "persisted";
-    }
-    const existing = await store.lookup(runId);
-    if (!existing) {
-      return "error";
-    }
-    if (existing.status === "cleanup") {
-      return "terminal";
-    }
-    return existing.key === value.key ? "persisted" : "conflict";
-  } catch (error) {
-    logFailure(api, "state store write", error);
-    return "error";
-  }
-}
-
-function persistedProgressRunFromTracker(
-  tracker: ProgressTracker,
-  status: PersistedProgressRun["status"],
-): PersistedProgressRun {
-  return {
-    key: `${tracker.accountId}:${tracker.channelId}:${tracker.messageId}`,
-    accountId: tracker.accountId,
-    channelId: tracker.channelId,
-    messageId: tracker.messageId,
-    status,
-    ...(tracker.runningEmoji ? { runningEmoji: tracker.runningEmoji } : {}),
-  };
-}
-
-async function markProgressRunForCleanup(
-  api: ProgressApi,
-  runId: string,
-  persisted: PersistedProgressRun,
-) {
-  try {
-    await getProgressStore(api)?.register(runId, { ...persisted, status: "cleanup" });
-    return true;
-  } catch (error) {
-    logFailure(api, "state store cleanup mark", error);
-    return false;
-  }
-}
-
-async function lookupProgressRun(
-  api: ProgressApi,
-  runId: string,
-): Promise<ProgressRunLookupResult> {
-  const store = getProgressStore(api);
-  if (!store) {
-    return { status: "error" };
-  }
-  try {
-    const value = await store.lookup(runId);
-    return value ? { status: "found", value } : { status: "missing" };
-  } catch (error) {
-    logFailure(api, "state store read", error);
-    return { status: "error" };
-  }
-}
-
-async function consumeProgressRun(api: ProgressApi, runId: string) {
-  try {
-    return await getProgressStore(api)?.consume(runId);
-  } catch (error) {
-    logFailure(api, "state store consume", error);
-    return undefined;
-  }
-}
-
-async function listProgressStateForKey(
-  api: ProgressApi,
-  key: string,
-): Promise<ProgressStateForKeyResult> {
-  const store = getProgressStore(api);
-  if (!store) {
-    return { ok: false };
-  }
-  try {
-    const entries = await store.entries();
-    const matching = entries.filter((entry) => entry.value.key === key);
-    return {
-      ok: true,
-      activeRunIds: matching
-        .filter((entry) => entry.value.status === "active")
-        .map((entry) => entry.key),
-      cleanupRunIds: matching
-        .filter((entry) => entry.value.status === "cleanup")
-        .map((entry) => entry.key),
-      ownedEmojis: Array.from(new Set(matching.flatMap((entry) => entry.value.runningEmoji ?? []))),
-    };
-  } catch (error) {
-    logFailure(api, "state store list", error);
-    return { ok: false };
-  }
-}
-
-async function runQueued(key: string, task: () => Promise<void>) {
-  const previous = trackerQueues.get(key) ?? Promise.resolve();
-  const current = previous.catch(() => undefined).then(task);
-  trackerQueues.set(key, current);
-  try {
-    await current;
-  } finally {
-    if (trackerQueues.get(key) === current) {
-      trackerQueues.delete(key);
-    }
-  }
-}
-
-function logFailure(api: ProgressApi, action: string, error: unknown) {
-  const message = error instanceof Error ? error.message : String(error);
-  api.logger.debug?.(`discord subagent progress ${action} failed: ${message}`);
 }
 
 async function setReaction(api: ProgressApi, tracker: ProgressTracker, emoji: string) {
@@ -519,7 +272,7 @@ async function handleStarted(
   event: Extract<SubagentProgressEvent, { phase: "started" }>,
 ) {
   const runId = event.runId.trim();
-  const target = resolveTarget(event);
+  const target = resolveDiscordProgressTarget(event.requester);
   if (!runId || !target || isRunTerminal(runId)) {
     return;
   }
@@ -728,7 +481,7 @@ function scheduleTerminalLookupRetry(
     return;
   }
   if (!owned) {
-    const target = resolveTarget({ phase: "started", runId, requester: event.requester });
+    const target = resolveDiscordProgressTarget(event.requester);
     const account = resolveDiscordAccount({
       cfg: api.config,
       accountId: event.requester?.accountId,
@@ -942,7 +695,6 @@ export function resetDiscordSubagentProgressForTest() {
   }
   trackers.clear();
   trackerKeyByRunId.clear();
-  trackerQueues.clear();
   terminalRunIds.clear();
   for (const timer of terminalRetryTimers.values()) {
     clearTimeout(timer);
@@ -956,5 +708,5 @@ export function resetDiscordSubagentProgressForTest() {
     }
   }
   startupRecoveryRetries.clear();
-  progressStores = new WeakMap();
+  resetDiscordSubagentProgressStateForTest();
 }
